WIP: update CryptoSSEService mock SSE for testing

+++ StockDashboard/app/src/main/java/vn/edu/usth/stockdashboard/data/model/StockItem.java
@@ -23,7 +23,7 @@ public class StockItem {
 
     private double investedValue; // Giá trị đã đầu tư
     private double currentValue;  // Giá trị hiện tại của khoản đầu tư
-    private int quantity; // Số lượng cổ phiếu nắm giữ
+    private double quantity; // Số lượng cổ phiếu nắm giữ
 
 
     // Constructor for creating initial placeholders
@@ -117,11 +117,11 @@ public class StockItem {
         return currentValue;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
     // Setters để cập nhật Portfolio Data
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
